backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["kjpar0317 <kjpar_kr@yahoo.co.kr>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.75.2"
uvicorn = "^0.17.6"
requests = "^2.27.1"
python-dotenv = "^0.20.0"
pydantic = "^1.9.0"
googletrans = "3.1.0a0"
autopep8 = "^1.6.0"
numpy = "^1.22.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
